added Java-Script example for allocating a WASM buffer.

+++ crates/svm-codec/src/lib.rs
@@ -47,7 +47,7 @@ use version::{decode_version, encode_version};
 /// If the encoding failed, the returned WASM buffer will contain a JSON with the error.
 #[no_mangle]
 #[cfg(target_arch = "wasm32")]
-pub extern "C" fn wasm_deploy_template(buf_ptr: i64) -> i64 {
+pub extern "C" fn wasm_deploy_template(buf_ptr: i32) -> i32 {
     let tx_ptr = api::wasm::encode_deploy_template(buf_ptr as usize);
 
     tx_ptr as _
@@ -62,7 +62,7 @@ pub extern "C" fn wasm_deploy_template(buf_ptr: i64) -> i64 {
 /// If the encoding failed, the returned WASM buffer will contain a JSON with the error.
 #[no_mangle]
 #[cfg(target_arch = "wasm32")]
-pub extern "C" fn wasm_spawn_app(buf_ptr: i64) -> i64 {
+pub extern "C" fn wasm_spawn_app(buf_ptr: i32) -> i32 {
     let tx_ptr = api::wasm::encode_spawn_app(buf_ptr as usize);
 
     tx_ptr as _
@@ -77,7 +77,7 @@ pub extern "C" fn wasm_spawn_app(buf_ptr: i64) -> i64 {
 /// If the encoding failed, the returned WASM buffer will contain a JSON with the error.
 #[no_mangle]
 #[cfg(target_arch = "wasm32")]
-pub extern "C" fn wasm_exec_app(buf_ptr: i64) -> i64 {
+pub extern "C" fn wasm_exec_app(buf_ptr: i32) -> i32 {
     let tx_ptr = api::wasm::encode_exec_app(buf_ptr as usize);
 
     tx_ptr as _
@@ -90,7 +90,7 @@ pub extern "C" fn wasm_exec_app(buf_ptr: i64) -> i64 {
 /// For more info read: `api::wasm::alloc`
 #[no_mangle]
 #[cfg(target_arch = "wasm32")]
-pub extern "C" fn wasm_alloc(length: i64) -> i64 {
+pub extern "C" fn wasm_alloc(length: i32) -> i32 {
     let ptr = api::wasm::alloc(length as usize);
 
     ptr as _
@@ -103,7 +103,7 @@ pub extern "C" fn wasm_alloc(length: i64) -> i64 {
 /// For more info read: `api::wasm::free`
 #[no_mangle]
 #[cfg(target_arch = "wasm32")]
-pub extern "C" fn wasm_free(buf_ptr: i64) {
+pub extern "C" fn wasm_free(buf_ptr: i32) {
     api::wasm::free(buf_ptr as usize);
 }
 
@@ -112,7 +112,7 @@ pub extern "C" fn wasm_free(buf_ptr: i64) {
 /// Returns the buffer `Data` byte-length
 #[no_mangle]
 #[cfg(target_arch = "wasm32")]
-pub extern "C" fn wasm_buffer_length(buf_ptr: i64) -> i32 {
+pub extern "C" fn wasm_buffer_length(buf_ptr: i32) -> i32 {
     let buf_len = api::wasm::wasm_buf_len(buf_ptr as usize);
 
     buf_len as _
@@ -123,7 +123,7 @@ pub extern "C" fn wasm_buffer_length(buf_ptr: i64) -> i32 {
 /// Returns a pointer to the buffer `Data`
 #[no_mangle]
 #[cfg(target_arch = "wasm32")]
-pub extern "C" fn wasm_buffer_data(buf_ptr: i64) -> i64 {
+pub extern "C" fn wasm_buffer_data(buf_ptr: i32) -> i32 {
     let (data_ptr, _len) = api::wasm::wasm_buf_data_ptr(buf_ptr as usize);
 
     data_ptr as _
